# src/main/resources/api/swagger.yaml
...
swagger: "2.0"
info:
  description: The Book API is a generic, flexible, configurable endpoint which allows
    requesting information on one or more books.
  version: 1.0.0
  title: Bookshop API
  termsOfService: http://www.termsofservice.org
  contact:
    name: Konstantin Priluchnyi
    url: https://github.com/konstantin-evo
    email: konstantin.priluchnyi@gmail.com
  license:
    name: License of API
    url: API license URL
basePath: /bookshop/1.0.0
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
...
paths:
  "/api/authors/{slug}":
    get:
      tags:
        - author-api-controller
      summary: getAuthorBySlug
      operationId: getAuthorBySlugUsingGET
      parameters:
        - name: slug
          in: path
          description: slug
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/AuthorDto"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    patch:
      tags:
        - author-api-controller
      summary: editAuthor
      operationId: editAuthorUsingPATCH
      parameters:
        - in: body
          name: authorDto
          description: authorDto
          required: true
          schema:
            "$ref": "#/definitions/AuthorDto"
        - name: slug
          in: path
          description: slug
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/ValidatedResponseDto"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  "/api/books/author/{id}":
    get:
      tags:
        - book-api-controller
      summary: getBooksByAuthor
      operationId: getBooksByAuthorUsingGET
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
        - name: limit
          in: query
          description: limit
          required: true
          type: integer
          format: int32
        - name: offset
          in: query
          description: offset
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/BooksPageDto"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
definitions:
...
  AuthorDto:
    type: object
    properties:
      description:
        type: object
        example: Dostoevsky's paternal ancestors were part of a noble family of Russian
          Orthodox Christians.The family traced its roots back to Danilo Irtishch,
          who was granted lands in the Pinsk region(for centuries part of the Grand
          Duchy of Lithuania, now in modern-day Belarus)in 1509 for his services under
          a local prince, his progeny then taking the name "Dostoevsky"based on a
          village there called Dostoïevo (derived from Old Polish dostojnik – dignitary).
        description: author's life story
      firstName:
        type: object
        example: Fyodor
        description: personal name of the author
      id:
        type: object
        example: 346
        description: uniq identifier
      lastName:
        type: object
        example: Dostoevsky
        description: family name of the author
    title: AuthorDto
    description: DTO Object for a person who writes book(s)
...
  BookDto:
    type: object
    properties:
      author:
        "$ref": "#/definitions/AuthorDto"
      description:
        type: string
        example: 'Raskolnikov, a destitute and desperate former student, wanders through
          the slums of St Petersburg and commits a random murder without remorse or
          regret. He imagines himself to be a great man, a Napoleon: acting for a
          higher purpose beyond conventional moral law.'
        description: book description text
      fileList:
        type: array
        example: modern, psychological prose
        description: tags added to the book
        items:
          "$ref": "#/definitions/BookFileDto"
      id:
        type: object
        example: 346
        description: uniq identifier
      isActive:
        type: integer
        format: int32
        example: 1
        description: if isActive = 1 so the book is considered to be active (possible
          to possibly buy, etc.)
      isBestseller:
        type: integer
        format: int32
        example: 1
        description: if isBestseller = 1 so the book is considered to be bestseller
          and if 0 the book is not a bestseller
      price:
        type: object
        example: 1250 ₽.
        description: book price at the moment without discounts and so on
      priceOld:
        type: string
        example: 1110 ₽.
        description: book price with discounts
      pubDate:
        type: string
        format: date
        example: '1866-03-20'
        description: date of book publication
      rating:
        type: integer
        format: int32
        example: 1280
        description: rating of the book
      tags:
        type: array
        example: modern, psychological prose
        description: tags added to the book
        uniqueItems: true
        items:
          "$ref": "#/definitions/TagDto"
      title:
        type: object
        example: Crime and Punishment
        description: the name of a book
    title: BookDto
    description: DTO Object for book entity
  BookFileDto:
    type: object
    properties:
      fileType:
        type: string
      hash:
        type: string
      path:
        type: string
    title: BookFileDto
  BooksPageDto:
    type: object
    properties:
      books:
        type: array
        items:
          "$ref": "#/definitions/BookDto"
      count:
        type: integer
        format: int32
    title: BooksPageDto
  TagDto:
    type: object
    properties:
      htmlTag:
        type: string
      id:
        type: integer
        format: int32
      name:
        type: string
    title: TagDto
  ValidatedResponseDto:
    type: object
    properties:
      errorMessages:
        type: object
        additionalProperties:
          type: string
      validated:
        type: boolean
    title: ValidatedResponseDto